% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AalborgWWTPs.R
\docType{AalborgWWTPs}
\name{AalborgWWTPs}
\alias{AalborgWWTPs}
\title{A subset of the MiDAS example data}
\format{An object of class \code{"ampvis2"}. See \code{\link[ampvis2]{amp_load}}.}
\source{
\href{http://midasfieldguide.org}{MiDAS: Field Guide to the Microbes of Activated Sludge and Anaerobic Digesters}
}
\usage{
data(AalborgWWTPs)
}
\description{
A smaller subset of \code{\link{MiDAS}} with 50 samples from 2 Danish Wastewater Treatment Plants; Aalborg West and Aalborg East.
}
\examples{
data(AalborgWWTPs)
AalborgWWTPs
}
\references{
Simon Jon McIlroy, Aaron Marc Saunders, Mads Albertsen, Marta Nierychlo, Bianca McIlroy, Aviaja Anna Hansen, Søren Michael Karst, Jeppe Lund Nielsen, Per Halkjær Nielsen; MiDAS: the field guide to the microbes of activated sludge, Database, Volume 2015, 1 January 2015, bav062
(\href{https://doi.org/10.1093/database/bav062}{https://doi.org/10.1093/database/bav062})
}
\keyword{datasets}
