% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{abundAreCounts}
\alias{abundAreCounts}
\title{Check whether abundances appear to be read counts (i.e. not normalised)}
\usage{
abundAreCounts(data)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}
}
\value{
Logical
}
\description{
Check whether abundances appear to be read counts (i.e. not normalised)
}
\keyword{internal}
