% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_alpha.R
\name{amp_alpha}
\alias{amp_alpha}
\title{Calculate alpha-diversity statistics for each sample}
\usage{
amp_alpha(data)
}
\arguments{
\item{data}{(required) An ampvis object.}

\item{measure}{Alpha-diversity measures to be included (default:Observed).}
}
\description{
Calculate alpha-diversity statistics for each sample and combine it with metadata.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
