% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_alphadiv.R
\name{amp_alphadiv}
\alias{amp_alphadiv}
\title{Alpha-diversity analysis}
\usage{
amp_alphadiv(data, measure = c(""), rarefy = 10000)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{amp_load()}.}

\item{measure}{Alpha-diversity measure(s) to be included if not all. A vector of one or more of: 
\itemize{
  \item \code{"observed"}
  \item \code{"shannon"}
  \item \code{"simpson"}
  \item \code{"invsimpson"}
}}

\item{rarefy}{Rarefy species richness to this value. (\emph{default:} \code{10000})}
}
\value{
A data frame.
}
\description{
Calculates alpha-diversity statistics for each sample.
}
\details{
\code{amp_alphadiv} simply calculates alpha-diversity indices using the vegan function \code{\link[vegan]{diversity}} and returns the results combined with the metadata from the provided ampvis2 object. For more details about the exact formulas used see the references.
}
\examples{

}
\references{
\url{https://cran.r-project.org/web/packages/vegan/vignettes/diversity-vegan.pdf}
  
  Hill, M. (1973). "Diversity and Evenness: A Unifying Notation and Its Consequences". Ecology, 54(2), 427-432. \url{doi.org/10.2307/1934352}
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
