% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_core.R
\name{amp_core}
\alias{amp_core}
\title{Generates a ggplot2 style core community plot}
\usage{
amp_core(data)
}
\arguments{
\item{data}{(required) A ampvis object.}

\item{group}{Group the data based on a sample variable (default: "Sample").}

\item{abund.treshold}{Treshold for considering something abundant in percent (default: 0.1).}

\item{tax.aggregate}{The taxonomic level that the data should be aggregated to (default: OTU).}

\item{tax.class}{Converts a specific phyla to class level instead (e.g. "p__Proteobacteria").}

\item{tax.empty}{Either "remove" OTUs without taxonomic information, add "best" classification or add the "OTU" name (default: best).}

\item{plotly}{Returns an interactive plot instead (default: F).}

\item{raw}{Display raw input instead of converting to percentages (default: F).}

\item{output}{Either plot or complete (default: "plot").}
}
\value{
A ggplot2 object
}
\description{
Generates a ggplot2 style core community plot
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
