% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_diffabund.R
\name{amp_diffabund}
\alias{amp_diffabund}
\title{Differential abundance test}
\usage{
amp_diffabund(data, group_by)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{amp_load()}.}

\item{group_by}{(required) The group of samples to test as defined by a categorical variable in the metadata.}

\item{test}{x}

\item{fitType}{x}

\item{num_threads}{x}

\item{signif_thrh}{Significance threshold. (\emph{default:} \code{0.01})}

\item{fold}{Log2fold filter for displaying significant results. (\emph{default:} \code{0})}

\item{label}{x}

\item{signif_plot_type}{Either \code{"boxplot"} or \code{"point"}. (\emph{default:} \code{"point"})}

\item{plot_nshow}{The amount of the most significant results to display. (\emph{default:} \code{10})}

\item{plot_point_size}{The size of the plotted points. (\emph{default:} \code{2})}

\item{tax_aggregate}{The taxonomic level to aggregate the OTUs. (\emph{default:} \code{"Phylum"})}

\item{tax_add}{Additional taxonomic level(s) to display, e.g. \code{"Phylum"}. (\emph{default:} \code{"none"})}

\item{tax_class}{Converts a specific phylum to class level instead, e.g. \code{"p__Proteobacteria"}.}

\item{tax_empty}{How to show OTUs without taxonomic information. One of the following:
\itemize{
   \item \code{"remove"}: Remove OTUs without taxonomic information.
   \item \code{"best"}: (\emph{default}) Use the best classification possible. 
   \item \code{"OTU"}: Display the OTU name.
   }}

\item{adjust_zero}{Keep 0 abundances in ggplot2 median calculations by adding a small constant to these.}

\item{plotly}{Returns an interactive plot instead. (\emph{default:} \code{FALSE})}
}
\value{
A list with multiple elements.
}
\description{
Tests if there is a significant difference in abundances between samples or groups hereof based on selected conditions.
}
\examples{
#Load example data
data("AalborgWWTPs")

#Save the results in an object
results <- amp_diffabund(AalborgWWTPs, group_by = "Plant", num_threads = 4)

#Show plots
results$plot_signif
results$plot_MA

#Or show raw results
results$DESeq2_results

}
\author{
Kasper Skytte Andersen \email{kasperskytteandersen@gmail.com}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
