% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_export_fasta.R
\name{amp_export_fasta}
\alias{amp_export_fasta}
\title{Export sequences from an ampvis object}
\usage{
amp_export_fasta(data)
}
\arguments{
\item{data}{(required) A data list loaded with amp_load.}

\item{file}{Name of the file containing the exported sequences.}

\item{tax}{Add taxonomic strings to the output (default: T).}
}
\description{
Export sequences from an ampvis object
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
