% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_export_otutable.R
\name{amp_export_otutable}
\alias{amp_export_otutable}
\title{Export otutable from an ampvis object}
\usage{
amp_export_otutable(data)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{amp_load()}.}

\item{file}{Name of the file containing the exported otutable.}

\item{id}{Name the samples using any variable in the metadata.}

\item{sort.samples}{Vector to sort the samples by.}

\item{raw}{Export raw input instead of converting to percentages (default: F)}
}
\description{
Export otutable from an ampcis object.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
