% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_function.R
\name{amp_function}
\alias{amp_function}
\title{Generate a function table}
\usage{
amp_function(heatmap)
}
\arguments{
\item{heatmap}{(required) A heatmap generated by amp_heatmap.}

\item{functions}{A data frame that associates functions with a genus names (default: "MiDAS").}

\item{adjust.bottom}{Adjust the whitespace added in the bottom of the plot in mm (default: 10).}

\item{adjust.top}{Adjust the whitespace added in the bottom of the plot in mm (default: 0).}

\item{genus.pos}{Specify which part of the name that contains the genus name (default: 1).}

\item{point.size}{Size of the plotted points (default: 6)}

\item{selection}{A vector with the functions to be displayed. (default: c("MiDAS","FIL", "AOB", "NOB", "PAO", "GAO").}
}
\value{
A ggplot2 object.
}
\description{
Generate a function table that is sorted based on an input heatmap.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
