% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_import_biom.R
\name{amp_import_biom}
\alias{amp_import_biom}
\title{Import OTU table from BIOM file}
\usage{
amp_import_biom(file)
}
\arguments{
\item{file}{Path to the .biom file.}
}
\value{
A data frame
}
\description{
Reads a .biom file and converts it into an OTU table compatible with ampvis2. This enables support for both the \href{http://qiime.org/}{QIIME} and \href{https://www.mothur.org/}{mothur} bioinformatic pipelines, as both software tools can output data in the \href{http://biom-format.org/}{BIOM format} (for mothur see \href{https://www.mothur.org/wiki/Make.biom}{make.biom}). Utilises the \href{https://github.com/joey711/biomformat}{biomformat} package, so both the JSON and HDF5 versions of the BIOM format are supported.
}
\examples{
\dontrun{
# First import the BIOM format OTU table:
biom_otutable <- amp_import_biom("path/to/file.biom")

# Then use amp_load() with or without metadata as normal:
d <- amp_load(biom_otutable, metadata)
}
}
\seealso{
\code{\link{amp_load}}, \code{\link{amp_import_usearch}}
}
