% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_load.R
\name{amp_load}
\alias{amp_load}
\title{Load data and convert to a list.}
\usage{
amp_load(otutable, metadata)
}
\arguments{
\item{otutable}{(required) A OTU table. The first row should be OTU name and the last 7 rows taxonomy.}

\item{metadata}{(required) A metadata file with sample names in first column.}

\item{refseq}{Reference sequences for all OTUs. Must be loaded with readDNAStringSet() from the biostrings package.}
}
\value{
A phyloseq object.
}
\description{
Load data and convert to a list.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
