% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_normalise.R
\name{amp_normalise}
\alias{amp_normalise}
\title{Normalise data.}
\usage{
amp_normalise(data)
}
\arguments{
\item{data}{(required) Data list as loaded with `amp_load()`.}
}
\value{
A list object.
}
\description{
Normalise data.
}
\author{
Rasmus Kirkegaard
}
