% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_octave.R
\name{amp_octave}
\alias{amp_octave}
\title{Octave plot}
\usage{
amp_octave(data, facet_by = "")
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{facet_by}{Split the plot into subplots based on a categorical/discrete variable in the samples metadata. (\emph{default:} \code{NULL})}

\item{scales}{If \code{facet_by} is set, should the axis scales of each subplot be fixed (\code{fixed}), free (\code{"free"}), or free in one dimension (\code{"free_x"} or \code{"free_y"})? (\emph{default:} \code{"fixed"})}

\item{num_threads}{Maximum number of distinct groups as defined by \code{facet_by} to process simultaneously using multicore processors. Only used if \code{facet_by} is set and there are more than one distinct group in the given variable(s). Defaults is the number of available cores minus 1.}
}
\value{
A ggplot2 object
}
\description{
Generates an octave plot to assess diversity. An octave plot is a histogram of the number of OTU's observed by bins of read counts, where the bin ranges increase exponentially, see details.
}
\details{
The \eqn{n}th bin in the histogram has the range \eqn{r(n)=2^n...2^{n+1}-1}.
The height of the bars then reflect the number of unique OTU's with read counts in each bin.
By judging the distribution one can assess whether the samples have been sequenced deeply
enough. A full symmetrical bell-shaped distribution with the left part far from the y-axis is the ideal.
A high amount of OTU's with a low amount of reads indicates noise, chimeras, and even cross talk.
}
\examples{
# Load example data
data("MiDAS")

# Generate an octave plot of all samples
amp_octave(MiDAS)

# Generate an octave plot for each group as
# defined by the metadata variable "Year", set the
# y-axes free and process 4 groups simultaneously using
# multicore processing
amp_octave(MiDAS,
  facet_by = "Year",
  scales = "free_y",
  num_threads = 4
)
}
\seealso{
\code{\link{amp_load}}, \code{\link{amp_rarecurve}}
}
\author{
Kasper Skytte Andersen \email{ksa@bio.aau.dk}
}
