% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_otu_network.R
\name{amp_otu_network}
\alias{amp_otu_network}
\title{Generate a taxa/sample network}
\usage{
amp_otu_network(data)
}
\arguments{
\item{data}{(required) Data list as loaded with `amp_load()`.}

\item{tax.aggregate}{The taxonomic level that the data should be aggregated to (default: "Phylum")}

\item{tax.add}{Additional taxonomic levels to display for each entry e.g. "Phylum" (default: "none")}

\item{tax.show}{The number of taxa to show (default: "all").}

\item{tax.class}{Converts a specific phyla to class level instead (e.g. "p__Proteobacteria").}

\item{tax.empty}{Either "remove" OTUs without taxonomic information, add "best" classification or add the "OTU" name (default: "best").}

\item{output}{To output a plot or the complete data inclusive dataframes (default: "plot")}

\item{raw}{Display raw input instead of converting to percentages (default: F)}

\item{min.abundance}{Minimum taxa abundance pr. sample (default: 0)}

\item{color}{A metadata variable to color the samples by.}
}
\value{
A ggplot2 object or a list with the ggplot2 object and associated dataframes.
}
\description{
Generate a taxa/sample network
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
