% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_rankabundance.R
\name{amp_rankabundance}
\alias{amp_rankabundance}
\title{Generates a ggplot2 style rank abundance plots}
\usage{
amp_rankabundance(data)
}
\arguments{
\item{data}{(required) A phyloseq object.}

\item{group}{Group the data based on a sample variable.}

\item{tax.clean}{Replace the phylum Proteobacteria with the respective Classes instead (default: T).}

\item{plot.log}{Log10 scale the data (default: F)}

\item{output}{Either "plot" or "complete" (default: "plot").}

\item{tax.add}{Additional taxonomic levels to display for each entry (default: "Phylum")}

\item{tax.aggregate}{The taxonomic level that the data should be aggregated to (defualt: "Genus")}

\item{tax.empty}{Either "remove" OTUs without taxonomic information, "rename" with best classification or add the "OTU" name (default: rename).}

\item{tax.class}{Converts a specific phyla to class level instead (e.g. "p__Proteobacteria").}

\item{raw}{Display raw input instead of converting to percentages (default: F).}

\item{order.group}{A vector defining the order of groups.}
}
\value{
A ggplot2 object
}
\description{
A nice long description
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
