% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_rarecurve.R
\name{amp_rarecurve}
\alias{amp_rarecurve}
\title{Rarefaction curve}
\usage{
amp_rarecurve(data)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{amp_load()}.}

\item{step}{Step size for sample sizes in rarefaction curves (default: 1000).}

\item{color}{A metadata variable to color by.}
}
\description{
Rarefaction curves for each sample
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
