% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{amp_rarefy}
\alias{amp_rarefy}
\title{Rarefy ampvis2 object}
\usage{
amp_rarefy(data, rarefy)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{rarefy}{(\emph{required}) Passed directly to \code{\link[vegan]{rrarefy}}.}
}
\value{
An ampvis2 object with rarefied OTU abundances.
}
\description{
This is just a wrapper of \code{\link[vegan]{rrarefy}} with convenient error messages and adjusted to work with ampvis2 objects.
}
\examples{
data("AalborgWWTPs")
AalborgWWTPs
rarefied <- amp_rarefy(AalborgWWTPs, 20000)
rarefied
}
\author{
Kasper Skytte Andersen \email{ksa@bio.aau.dk}
}
