% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_subset_samples.R
\name{amp_subset_samples}
\alias{amp_subset_samples}
\title{Subset data based on metadata.}
\usage{
amp_subset_samples(data, ...)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{amp_load()}.}

\item{...}{(optional) Additional arguments for the \code{subset()} function.}

\item{minreads}{(optional) Minimum number of reads pr. sample. (\emph{default:} \code{1})}
}
\value{
A list with 3 dataframes (4 if reference sequences are provided).
}
\description{
Subset data based on metadata.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
