% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_subset_taxa.R
\name{amp_subset_taxa}
\alias{amp_subset_taxa}
\title{Subset data based on taxonomy.}
\usage{
amp_subset_taxa(data, ...)
}
\arguments{
\item{data}{(required) A object.}

\item{...}{(required) Additional data is passed on to the R subset function. The samples can be subset based on any variable in the taxonomy.}
}
\value{
A phyloseq object.
}
\description{
Subset data based on taxonomy.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
