% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampvis2.R
\docType{package}
\name{ampvis2}
\alias{ampvis2}
\title{\code{ampvis2} package}
\description{
\code{ampvis2} is an R-package to conveniently visualise and analyse 16S rRNA amplicon data in different ways.
}
\details{
Please use the online documentation webpage at \href{https://madsalbertsen.github.io/ampvis2/}{https://madsalbertsen.github.io/ampvis2}
}
