% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{as.data.table.DNAbin}
\alias{as.data.table.DNAbin}
\title{S3 generic function to create a data table from a DNAbin class object}
\usage{
\method{as.data.table}{DNAbin}(x, ...)
}
\arguments{
\item{x}{(\emph{required}) A DNAbin class object}

\item{...}{Not used}
}
\value{
a data.table
}
\description{
S3 generic function to create a data table from a DNAbin class object
}
