% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{checkReqPkg}
\alias{checkReqPkg}
\title{Check for installed package}
\usage{
checkReqPkg(pkg, msg = "")
}
\arguments{
\item{pkg}{The package to check for, a character of length 1.}

\item{msg}{Optionally additional text appended (with \code{paste0}) to the default error message.}
}
\value{
Returns error and message if not installed, otherwise \code{invisible(TRUE)}
}
\description{
Returns error if a required package is not installed. Mostly used for checking whether packages listed under the Suggests field in the DESCRIPTION file is installed.
}
\author{
Kasper Skytte Andersen \email{ksa@bio.aau.dk}
}
\keyword{internal}
