% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{dist.JSD}
\alias{dist.JSD}
\title{Calculate Jensen-Shannon Divergence distances}
\usage{
dist.JSD(abund, pseudocount = 0.000001)
}
\arguments{
\item{abund}{Abundance table with OTU counts, in \code{ampvis2} objects it is available with simply data$abund.}

\item{pseudocount}{Pseudocount to use instead of 0-divisions. (\emph{default:} \code{0.000001})}
}
\value{
A distance matrix of class \code{dist}.
}
\description{
Calculate Jensen-Shannon Divergence distances
}
