% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_metadata}
\alias{example_metadata}
\title{A minimal example of sample metadata}
\format{a data frame.}
\description{
A minimal example of a metadata sheet that is compatible with \code{\link{amp_load}} and ampvis2 functions.
}
\examples{
\dontrun{
data(example_metadata)
View(example_metadata)
}
}
\keyword{data}
