% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_otutable}
\alias{example_otutable}
\title{A minimal example of an OTU-table}
\format{a data frame.}
\description{
A minimal example of an OTU-table that is compatible with \code{\link{amp_load}} and ampvis2 functions.
}
\examples{
\dontrun{
data(example_otutable)
View(example_otutable)
}
}
\keyword{data}
