% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_taxonomy}
\alias{example_taxonomy}
\title{A minimal example of taxonomy}
\format{
a data frame.
}
\description{
A minimal example of a taxonomy table that is compatible with \code{\link{amp_load}} and ampvis2 functions.
}
\examples{
\dontrun{
data(example_taxonomy)
View(example_taxonomy)
}
}
\keyword{data}
