% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{extractFunctions}
\alias{extractFunctions}
\title{Extract functional information about Genera from the MiDAS field guide}
\usage{
extractFunctions(FGList)
}
\arguments{
\item{FGList}{Data list obtained by \code{\link{getMiDASFGData}}.}
}
\value{
A data frame where each row is a Genus and each column is a "function".
}
\description{
Extract field related to properties and metabolism of all Genera from a list obtained by \code{\link{getMiDASFGData}} and return in a data frame.
}
\author{
Kasper Skytte Andersen \email{ksa@bio.aau.dk}
}
\keyword{internal}
