% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{filter_otus}
\alias{filter_otus}
\alias{filter_species}
\title{Filter OTUs by a threshold in percent}
\usage{
filter_otus(data, filter_otus = 0)

filter_species(data, filter_otus = 0)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{filter_otus}{Remove low abundant OTU's across all samples below this threshold in percent. (\emph{default}: \code{0})}
}
\value{
An ampvis2 object
}
\description{
Removes all OTUs that are not found with a higher relative abundance than the set threshold in percent in at least one sample.
}
\examples{
data("AalborgWWTPs")
AalborgWWTPs
filtered <- filter_otus(AalborgWWTPs, filter_otus = 0.1)
filtered
}
