% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{filter_species}
\alias{filter_species}
\title{Filter species by a threshold in percent}
\usage{
filter_species(data, filter_species = 0)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{filter_species}{Remove low abundant OTU's across all samples below this threshold in percent. (\emph{default}: \code{0})}
}
\value{
An ampvis2 object
}
\description{
Filter species by a threshold in percent
}
