% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_load.R
\name{findOTUcol}
\alias{findOTUcol}
\title{Detect OTU/ASV column in a data frame}
\usage{
findOTUcol(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
A data frame
}
\description{
Detect OTU/ASV column in a data frame
}
\keyword{internal}
