% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{getLowestTaxLvl}
\alias{getLowestTaxLvl}
\title{Find lowest taxonomic level}
\usage{
getLowestTaxLvl(tax, tax_aggregate = NULL, tax_add = NULL)
}
\arguments{
\item{tax}{The taxonomy table from an ampvis2 object (\code{ampvis2obj$tax})}

\item{tax_aggregate}{A character vector of one or more taxonomic levels (exactly as in the column names of \code{ampvis2obj$tax}), fx Genus or Species. (\emph{default:} \code{NULL})}

\item{tax_add}{A second character vector similar to \code{tax_aggregate}. (\emph{default:} \code{NULL})}
}
\value{
A length one character vector with the lowest taxonomic level
}
\description{
Finds the lowest taxonomic level of two given levels in \code{tax}. The hierarchic order of taxonomic levels is simply taken from the order of column names in \code{tax}
}
