% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{getMiDASFGData}
\alias{getMiDASFGData}
\title{Get data from the MiDAS field guide API}
\usage{
getMiDASFGData()
}
\value{
A list with all fields.
}
\description{
Gets all fields from the MiDAS field guide (\url{https://midasfieldguide.org}) returned in a list.
}
\author{
Kasper Skytte Andersen \email{ksa@bio.aau.dk}
}
