% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_load.R
\name{import}
\alias{import}
\title{Import any file format}
\usage{
import(x, ...)
}
\arguments{
\item{x}{File path or data frame.}

\item{...}{Additional arguments passed on to read_excel() and fread().}
}
\value{
A data frame
}
\description{
Excel, txt, CSV, TSV, BIOM, sintax. If already a data frame, will return as-is.
}
