% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{is_ampvis2}
\alias{is_ampvis2}
\title{Check if data has class "ampvis2"}
\usage{
is_ampvis2(data)
}
\arguments{
\item{data}{Object to check}
}
\value{
Returns nothing, only error if \code{class(data) != "ampvis"}.
}
\description{
Checks if the object is of class "ampvis2".
}
\author{
Kasper Skytte Andersen \email{ksa@bio.aau.dk}
}
