% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{matchOTUs}
\alias{matchOTUs}
\title{Rename OTU's by sequence matching with a FASTA file}
\usage{
matchOTUs(data, fasta, unmatched_prefix = "unmatched", rename_unmatched = TRUE)
}
\arguments{
\item{data}{data (\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{fasta}{Path to a FASTA file or a \code{DNAbin} class object with sequences whose names will be used as OTU names by exact matches (i.e. same length, 100% sequence identity). (\emph{default:} \code{NULL})}

\item{unmatched_prefix}{Prefix used to name any unmatched sequences when \code{refseq_names} is provided. An integer counting from 1 will be appended to this prefix, so for example the 123th unmatched sequence will be named \code{unmatched123}, and so on. (\emph{default:} \code{"unmatched"})}

\item{rename_unmatched}{Whether to rename any unmatched sequences or not when \code{refseq_names} is provided. (\emph{default:} \code{TRUE})}
}
\value{
An ampvis2 class object
}
\description{
Match and rename OTU's in an ampvis2 object by sequence to a FASTA file
}
