% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{midasfunctions}
\alias{midasfunctions}
\title{Functional information about microbes at Genus level}
\format{
a data frame.
}
\description{
Grabbed from midasfieldguide.org at 2020-12-01
}
\examples{
\dontrun{
data(midasfunctions)
View(midasfunctions)
}
}
\keyword{datasets}
