% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{midasfunctions_20201201}
\alias{midasfunctions_20201201}
\title{Functional information about microbes at Genus level}
\format{
a data frame.
}
\description{
Grabbed from midasfieldguide.org at 2020-12-01
}
\examples{
\dontrun{
data(midasfunctions_20201201)
View(midasfunctions_20201201)
}
}
\keyword{data}
