% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MiF}
\alias{MiF}
\title{Functional information of microbes in the MiDAS database}
\format{A data frame.
\describe{
  \item{POS}{Positive.}
  \item{NEG}{Negative.}
  \item{VAR}{Variable within species.}
  \item{NT}{Not tested.}
}}
\description{
A data frame with known/unknown functional characteristics of the Genera in the MiDAS database.
}
\keyword{data}
