% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_load.R
\name{parseTaxonomy}
\alias{parseTaxonomy}
\title{Extract/parse taxonomy from a data.frame}
\usage{
parseTaxonomy(x)
}
\arguments{
\item{x}{Data frame or file path}
}
\value{
A data frame
}
\description{
Extract/parse taxonomy from a data.frame
}
