% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_load.R
\name{phyloseq_to_ampvis2}
\alias{phyloseq_to_ampvis2}
\title{Convert a phyloseq-class object to an ampvis2-class object}
\usage{
phyloseq_to_ampvis2(physeq)
}
\arguments{
\item{physeq}{Phyloseq object to import}
}
\value{
A list of class \code{"ampvis2"} with 3 to 5 elements.
}
\description{
Does as the title says. The phyloseq package is required and must first be installed using the BiocManager package.
}
\examples{
\dontrun{
require("ampvis2")
require("phyloseq")

#print object summary
phyloseq_object

#convert
ampvis2_object <- phyloseq_to_ampvis2(phyloseq_object)

#print object summary
ampvis2_object
}
}
