% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ampvis2.R
\name{print.ampvis2}
\alias{print.ampvis2}
\title{Print ampvis data summary. Internal function.}
\usage{
\method{print}{ampvis2}(data)
}
\arguments{
\item{data}{}
}
\value{
Text output in console
}
\description{
Print ampvis data summary. Internal function.
}
