% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.coreplot}
\alias{print.coreplot}
\title{Prints amp_core plots if margin plots are also generated (internal function)}
\usage{
\method{print}{coreplot}(x, ...)
}
\arguments{
\item{x}{plot}

\item{...}{not used}
}
\description{
Prints amp_core plots if margin plots are also generated (internal function)
}
\author{
Kasper Skytte Andersen \email{ksa@bio.aau.dk}
}
