% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.figcaption}
\alias{print.figcaption}
\title{Print method for figure caption created by amp_ordinate}
\usage{
\method{print}{figcaption}(captionwithrefs)
}
\arguments{
\item{captionwithrefs}{Character vector with the caption}
}
\description{
Print method for figure caption created by amp_ordinate
}
\author{
Kasper Skytte Andersen \email{ksa@bio.aau.dk}
}
