% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\docType{data}
\name{tax_levels}
\alias{tax_levels}
\title{Valid taxonomic levels}
\format{
A charactor vector with the taxonomic levels Kingdom->OTU
}
\usage{
tax_levels
}
\description{
Valid taxonomic levels
}
\keyword{data}
\keyword{internal}
