% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{unzip_file}
\alias{unzip_file}
\title{Unzip file}
\usage{
unzip_file(file)
}
\arguments{
\item{file}{data (\emph{required}) Data list as loaded with \code{\link{amp_load}}.}
}
\value{
Path to the decompressed file. If not a zip file, returns \code{file} without doing anything.
}
\description{
If the specified file has zip signature will unzip it to a tempfile and return the path to the decompressed file. The extension is kept from the file within the archive and only one file inside the archive is allowed.
}
\keyword{internal}
