% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AalborgWWTPs}
\alias{AalborgWWTPs}
\title{A subset of the MiDAS example data}
\format{
An object of class \code{"ampvis2"}. See \code{\link[ampvis2]{amp_load}}.
}
\description{
A smaller subset of \code{\link{MiDAS}} with 50 samples from 2 Danish Wastewater Treatment Plants; Aalborg West and Aalborg East.
}
\examples{
\dontrun{
data("AalborgWWTPs")
View(AalborgWWTPs)
}
}
\keyword{datasets}
