% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_alpha.R
\name{amp_alpha}
\alias{amp_alpha}
\title{Calculate alpha-diversity statistics}
\usage{
amp_alpha(data)
}
\arguments{
\item{data}{(required) Data list as loaded with `amp_load()`.}

\item{measure}{Alpha-diversity measures to be included. (default: `"Observed"`)}
}
\value{
A data frame
}
\description{
Calculates alpha-diversity statistics for each sample and combines with metadata.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
