% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_alpha.R
\name{amp_alpha}
\alias{amp_alpha}
\title{Calculate alpha-diversity statistics}
\usage{
amp_alpha(data, measure = "", rarefy = 10000)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{amp_load()}.}

\item{measure}{Alpha-diversity measure(s) to be included if not all. A vector of one or more of: \code{"observed"}, \code{"shannon"}, \code{"simpson"} or \code{"invsimpson"}.}

\item{rarefy}{Rarefy species richness to this value. (\emph{default:} \code{10000})}
}
\value{
A data frame
}
\description{
Calculates alpha-diversity statistics for each sample and combines with the metadata.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
