% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_alphadiv.R
\name{amp_alphadiv}
\alias{amp_alphadiv}
\title{Alpha-diversity analysis}
\usage{
amp_alphadiv(data, measure, rarefy)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{measure}{Alpha-diversity measure(s) to be included if not all. A vector of one or more of: 
\itemize{
  \item \code{"observed"}
  \item \code{"shannon"}
  \item \code{"simpson"}
  \item \code{"invsimpson"}
}}

\item{richness}{(\emph{logical}) Also calculate sample richness estimates (Chao1 and ACE) as calculated by \code{\link[vegan]{estimateR}}. (\emph{default:} \code{FALSE})}

\item{rarefy}{Rarefy species richness to this value before calculating alpha diversity and/or richness. Passed directly as the \code{sample} argument to \code{\link[vegan]{rrarefy}}. (\emph{default:} \code{NULL})}
}
\value{
A data frame.
}
\description{
Calculate alpha-diversity indices for each sample and combines with the metadata.
}
\details{
The alpha-diversity indices are calculated per sample using the vegan function \code{\link[vegan]{diversity}}, where the read abundances are first rarefied using \code{\link[vegan]{rrarefy}} by the size of the \code{rarefy} argument. Refer to the vegan documentation for details about the different indices and how they are calculated. If no measure(s) are chosen, all diversity indices will be returned.
}
\examples{
#Load example data
data("AalborgWWTPs")

#Subsample/rarefy to 20000 reads and then calculate Shannon and Simpson alpha-diversity indices
alphadiversityresult <- amp_alphadiv(AalborgWWTPs, measure = c("shannon", "simpson"), rarefy = 20000)

#Explore the results in the data frame
#View(alphadiversityresult)

}
\author{
Kasper Skytte Andersen \email{kasperskytteandersen@gmail.com}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
