% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_boxplot.R
\name{amp_boxplot}
\alias{amp_boxplot}
\title{Boxplot}
\usage{
amp_boxplot(data)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{group_by}{Group the samples by a variable in the metadata.}

\item{sort_by}{Sort the boxplots by \code{"median"}, \code{"mean"} or \code{"total"}. (\emph{default:} \code{"median"})}

\item{plot_type}{Plot type. \code{"boxplot"} or \code{"point"}. (\emph{default:} \code{"boxplot"})}

\item{point_size}{The size of points. (\emph{default:} \code{1})}

\item{tax_aggregate}{The taxonomic level to aggregate the OTUs. (\emph{default:} \code{"Genus"})}

\item{tax_add}{Additional taxonomic level(s) to display, e.g. \code{"Phylum"}. (\emph{default:} \code{"none"})}

\item{tax_show}{The number of taxa to show, or a vector of taxa names. (\emph{default:} \code{20})}

\item{tax_empty}{How to show OTUs without taxonomic information. One of the following:
\itemize{
   \item \code{"remove"}: Remove OTUs without taxonomic information.
   \item \code{"best"}: (\emph{default}) Use the best classification possible. 
   \item \code{"OTU"}: Display the OTU name.
   }}

\item{tax_class}{Converts a specific phylum to class level instead, e.g. \code{"p__Proteobacteria"}.}

\item{order_group}{A vector to order the groups by.}

\item{order_y}{A vector to order the y-axis by.}

\item{plot_flip}{(\emph{logical}) Flip the axes of the plot axis. (\emph{default:} \code{FALSE})}

\item{plot_log}{(\emph{logical}) Log10-scale the plot. (\emph{default:} \code{FALSE})}

\item{adjust_zero}{Keep abundances of 0 in the calculation of medians by adding this value. (\emph{default:} \code{NULL})}

\item{raw}{(\emph{logical}) Display raw input instead of converting to percentages. (\emph{default:} \code{FALSE})}

\item{detailed_output}{(\emph{logical}) Return additional details or not. If \code{TRUE}, it is recommended to save to an object and then access the additional data by \code{View(object$data)}. (\emph{default:} \code{FALSE})}
}
\value{
A ggplot2 object. If \code{detailed_output = TRUE} a list with a ggplot2 object and additional data.
}
\description{
Generates boxplots of the most abundant taxa.
}
\examples{
#Load example data
data("AalborgWWTPs")

#10 boxplots grouped by WWTP with phylum name added
amp_boxplot(AalborgWWTPs, 
            group_by = "Plant",
            tax_show = 10,
            tax_add = "Phylum")

}
\author{
Kasper Skytte Andersen \email{kasperskytteandersen@gmail.com}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
