% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_cleanMiF.R
\name{amp_cleanMiF}
\alias{amp_cleanMiF}
\title{Make raw MiDAS function data into ampvis format}
\usage{
amp_cleanMiF(data)
}
\arguments{
\item{data}{(required) A data frame with MiDAS functions.}
}
\description{
Make raw MiDAS function data into ampvis format. Internal function.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
