% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_core.R
\name{amp_core}
\alias{amp_core}
\title{Core community analysis}
\usage{
amp_core(data)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{amp_load()}.}

\item{group_by}{Group the samples by a variable in the metadata. (\emph{default:} \code{"Sample"})}

\item{abund_treshold}{Threshold in percent for defining "abundant"/"core" taxa. (\emph{default:} \code{0.1})}

\item{tax_aggregate}{The taxonomic level to aggregate the OTUs. (\emph{default:} \code{"OTU"})}

\item{tax_class}{Converts a specific phylum to class level instead, e.g. \code{"p__Proteobacteria"}.}

\item{tax_empty}{How to show OTUs without taxonomic information. One of the following:
\itemize{
   \item \code{"remove"}: Remove OTUs without taxonomic information.
   \item \code{"best"}: (\emph{default}) Use the best classification possible. 
   \item \code{"OTU"}: Display the OTU name.
   }}

\item{plotly}{(\emph{logical}) Returns an interactive plot instead. (\emph{default:} \code{FALSE})}

\item{raw}{(\emph{logical}) Display raw input instead of converting to percentages. (\emph{default:} \code{FALSE})}

\item{detailed_output}{(\emph{logical}) Return additional details or not. If \code{TRUE}, it is recommended to save to an object and then access the additional data by \code{View(object$data)}. (\emph{default:} \code{FALSE})}
}
\value{
A ggplot2 object. If \code{detailed_output = TRUE} a list with a ggplot2 object and additional data.
}
\description{
Generates a ggplot2 style core community plot of amplicon data
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
