% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_export_fasta.R
\name{amp_export_fasta}
\alias{amp_export_fasta}
\title{Export raw DNA sequences}
\usage{
amp_export_fasta(data)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{filename}{File name of the exported FASTA file. (\emph{default:} \code{"exported_sequences.fa"})}

\item{tax}{(\emph{logical}) Add taxonomic strings to the output or not. (\emph{default:} \code{TRUE})}
}
\description{
Export sequences from an ampvis2 object with added taxonomy.
}
\examples{
#Load example data
data("AalborgWWTPs")

#Export fasta sequences as "exported_sequences.fa"
\dontrun{
amp_export_fasta(AalborgWWTPs)
}

}
\author{
Kasper Skytte Andersen \email{kasperskytteandersen@gmail.com}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
