% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_export_otutable.R
\name{amp_export_otutable}
\alias{amp_export_otutable}
\title{Export OTU-table}
\usage{
amp_export_otutable(data)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{amp_load()}.}

\item{filename}{File name of the exported OTU-table. (\emph{default:} \code{"exported_otutable.txt"})}

\item{id}{Name the samples using a variable in the metadata.}

\item{sort_samples}{Vector to sort the samples by.}

\item{raw}{(\emph{logical}) Use raw counts instead of percentages. (\emph{default:} \code{FALSE})}
}
\description{
Export otutable (including taxonomy) from an ampvis2 object.
}
\examples{
#Load example data
data("AalborgWWTPs")

#Export OTU-table
\dontrun{
amp_export_otutable(AalborgWWTPs)
}

}
\author{
Kasper Skytte Andersen \email{kasperskytteandersen@gmail.com}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
