% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_export_otutable.R
\name{amp_export_otutable}
\alias{amp_export_otutable}
\title{Export OTU-table}
\usage{
amp_export_otutable(data)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{filename}{File name of the exported OTU-table WITHOUT extension. (\emph{default:} \code{"exported_otutable"})}

\item{md5}{(\emph{logical}) Compute md5 sum of the data (the whole object, not just otutable) and append to the filename. (\emph{default:} \code{FALSE})}

\item{sep}{Separator passed directly to \code{\link{write.table}}. (\emph{default:} \code{"\t"})}

\item{id}{Name the samples using a variable in the metadata.}

\item{sort_samples}{Vector to sort the samples by.}

\item{raw}{(\emph{logical}) Use raw counts instead of percentages. (\emph{default:} \code{TRUE})}

\item{...}{Additional arguments passed to \code{\link{write.table}} other than \code{sep} and \code{row.names}.}
}
\description{
Export otutable (including taxonomy) from an ampvis2 object as CSV using \code{\link{write.table}}.
}
\examples{
#Load example data
data("AalborgWWTPs")

#Export OTU-table
\dontrun{
amp_export_otutable(AalborgWWTPs, md5 = TRUE, filename = "AalborgWWTPs_otutable", sep = "\\t")
}

}
\author{
Kasper Skytte Andersen \email{kasperskytteandersen@gmail.com}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
