% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_function.R
\name{amp_function}
\alias{amp_function}
\title{Add functional information of taxa to a heatmap}
\usage{
amp_function(heatmap)
}
\arguments{
\item{heatmap}{(required) A heatmap as generated by \code{amp_heatmap()}.}

\item{function_data}{A data frame with functional information about genus-level OTUs. If none provided, the \code{data("MiF")} dataset will be used.}

\item{adjust_bottom}{Adjust the whitespace at the bottom of the plot in mm. (\emph{default:} \code{10})}

\item{adjust_top}{Adjust the whitespace at the top of the plot in mm. (\emph{default:} \code{0})}

\item{genus_pos}{Specify which part of the name that contains the genus name. (\emph{default:} \code{1})}

\item{point_size}{Size of the plotted points. (\emph{default:} \code{6})}

\item{plotted_functions}{A vector with the functions to be displayed. (\emph{default:} \code{c("MiDAS","FIL", "AOB", "NOB", "PAO", "GAO")})}
}
\value{
A ggplot2 object.
}
\description{
Generates a table with functional information about the taxa in an input heatmap and combines the two.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}

Kasper Skytte Andersen \email{ksa@bio.aau.dk}
}
