% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_import_usearch.R
\name{amp_import_usearch}
\alias{amp_import_usearch}
\title{Import OTU table from USEARCH pipelines}
\usage{
amp_import_usearch(otutab, sintax)
}
\arguments{
\item{otutab}{Path to the tabulated OTU count table as created by \code{usearch10 -otutab ... -otutabout} commands.}

\item{sintax}{Path to the tabulated sintax taxonomy table as created by \code{usearch10 -sintax ... -tabbedout} commands.}
}
\value{
A data frame
}
\description{
Reads an OTU count table and a corresponding sintax taxonomy table from a \href{http://www.drive5.com/usearch/}{USEARCH} analysis \href{http://www.drive5.com/usearch/manual/ex_miseq.html}{pipeline}, and then converts and combines them into an OTU table compatible with ampvis2.
}
\examples{
\dontrun{
# First import the usearch format OTU table:
usearch_otutable <- amp_import_usearch(otutab = "/path/to/otutab.txt", sintax = "/path/to/sintax.txt")

# Then use amp_load() with or without metadata as normal:
d <- amp_load(usearch_otutable, metadata)
}
}
