% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_merge_ampvis2.R
\name{amp_merge_ampvis2}
\alias{amp_merge_ampvis2}
\title{Merge ampvis2 object(s)}
\usage{
amp_merge_ampvis2(...)
}
\arguments{
\item{...}{Any number of ampvis2-class objects to merge}
}
\value{
An ampvis2-class object
}
\description{
Merge any number of ampvis2 objects into a single object
}
\details{
Use with care. It is the users responsibility alone to ensure that OTU ID's are not arbitrary and that they are corresponding to the same sequences across data sets (objects). The same applies to the taxonomy, which must be generated in the exact same way with the same database across data sets.
}
\examples{
data("MiDAS")

#summary of samples from 2010-2012
amp_subset_samples(
  MiDAS,
  Year \%in\% c("2010", "2011", "2012")
)

#now merge individual objects and verify summary is the same
d_2010 <- amp_subset_samples(
  MiDAS,
  Year \%in\% "2010"
)
d_2011 <- amp_subset_samples(
  MiDAS,
  Year \%in\% "2011"
)
d_2012 <- amp_subset_samples(
  MiDAS,
  Year \%in\% "2012"
)

amp_merge_ampvis2(
  d_2010,
  d_2011,
  d_2012
)
}
