% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_ordinate.R
\name{amp_ordinate}
\alias{amp_ordinate}
\title{Generates a ggplot2 style ordinate plot of amplicon data}
\usage{
amp_ordinate(data)
}
\arguments{
\item{data}{(required) Data list loaded with amp_load() containing the elements: abund, metadata and tax.}

\item{filter_species}{Remove low abundant OTU's across all samples below this threshold in percent. Recommended minimum: 0.1 pct (default: 0.1).}

\item{type}{Ordination type; Principal Components Analysis(PCA), Redundancy Analysis(RDA), non-metric Multidimensional Scaling(NMDS), metric Multidimensional Scaling(MMDS, aka PCoA), Correspondence Analysis(CA), Canonical Correspondence Analysis(CCA) or Detrended Correspondence Analysis(DCA) (Default: PCA)}

\item{metric}{Distance metric used for the distance-based ordination methods (nMDS/PCoA), any of the following: "manhattan", "euclidean", "canberra", "bray", "kulczynski", "jaccard", "gower", "jsd" (Jensen-Shannon Divergence), "altGower", "morisita", "horn", "mountford", "raup" , "binomial", "chao", "cao", "mahalanobis" or simply "none" or "sqrt". See details in ?vegdist, and for JSD http://enterotype.embl.de/enterotypes.html.}

\item{transform}{Transform the abundance table with the decostand() function, fx "normalize", "chi.square", "hellinger" or "sqrt", see details in ?decostand. Using the hellinger transformation is always a good choice and is recommended for PCA/RDA/nMDS/PCoA to obtain a more ecologically meaningful result (learn about the double-zero problem).}

\item{constrain}{Variable(s) in the metadata for constrained analyses (RDA and CCA). Multiple variables can be provided by a vector, fx c("Year", "Temperature"), but keep in mind the more variables the more the result will be similar to unconstrained analysis.}

\item{x_axis}{Which axis from the ordination results to plot as the first axis. Have a look at the $screeplot with output = "detailed" to validate axes (default: 1).}

\item{y_axis}{Which axis from the ordination results to plot as the second axis. Have a look at the $screeplot with output = "detailed" to validate axes (default: 2).}

\item{sample_color}{Color sample points by a variable in the metadata.}

\item{sample_color_order}{Order the colors in color by a vector.}

\item{sample_shape}{Shape sample points by a variable in the metadata.}

\item{sample_colorframe}{Frame the points with a polygon colored by the color argument (default: F).}

\item{sample_colorframe_label}{Label by a variable in the metadata.}

\item{sample_label}{Label sample points by a variable in the metadata.}

\item{sample_label_size}{Sample labels text size (default: 4).}

\item{sample_label_segment_color}{Sample labels repel-segment color (default: "black").}

\item{sample_trajectory}{Make a trajectory between sample points by a variable in the metadata.}

\item{sample_trajectory_group}{Make a trajectory between sample points by the trajectory argument, but within individual groups.}

\item{sample_plotly}{Enable sample point howevering to display specific metadata or "all" (default: none).}

\item{species_plot}{Plot species points (default: F).}

\item{species_nlabels}{Number of most extreme species labels to plot.}

\item{species_label_taxonomy}{Taxonomic level by which to label the species points (default: "Genus").}

\item{species_label_size}{Size of the species text labels (default: 3).}

\item{species_label_color}{Color of the species text labels (default: "grey10).}

\item{species_rescale}{rescale species (default: F).}

\item{species_size}{Size of the species points (default: 2).}

\item{species_shape}{The shape of the points, fx 1 for hollow circles or 20 for dots (default: 20).}

\item{species_plotly}{Enable species specific point howevering to display taxonomy (default: F).}

\item{envfit_factor}{A vector of factor variables from the sample data used for envfit to the model}

\item{envfit_numeric}{A vector of numerical variables from the sample data used for envfit to the model}

\item{envfit_signif_level}{The significance treshold for displaying envfit parameters (default: 0.001).}

\item{envfit_textsize}{Size of the envfit text on the plot (default: 3).}

\item{envfit_color}{Color of the envfit text on the plot (default: "darkred").}

\item{envfit_numeric_arrows_scale}{Scale the size of the numeric arrows (default: 1).}

\item{envfit_show}{Show the results on the plot (default: T).}

\item{repel}{Repel all labels to prevent cluttering of plots (default: T).}

\item{opacity}{Opacity of all plotted points and sample_colorframe opacity, 0:invisible, 1:opaque (default: 0.8).}

\item{tax.empty}{Option to add "best" classification or just the "OTU" name to each "OTU" (default: best).}

\item{output}{"plot" or "detailed"; output as list with additional information(model, scores, inputmatrix etc) or just the plot (default: "plot).}

\item{...}{Pass additional arguments to the vegan ordination functions, fx rda(...), cca(...), metaMDS(...), see vegan help.}
}
\value{
A ggplot2 object or a list with the ggplot2 object and associated dataframes.
}
\description{
A wrapper around the vegan package to make beautiful ggplot2 ordination plots to analyse and compare microbial community compositions. The input data must be a list loaded with amp_load() from the ampvis package, but any OTU table-like matrix (OTU's in rows, sampleID's in columns and abundances in the corresponding cells, aka a contingency table) alongside metadata for the same samples can be used. Simply choose an ordination type and a plot is returned.
}
\author{
Kasper Skytte Andersen \email{kasperskytteandersen@gmail.com}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
