% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_rankabundance.R
\name{amp_rankabundance}
\alias{amp_rankabundance}
\title{Rank abundance plot}
\usage{
amp_rankabundance(data, group_by, showSD = TRUE, log10_x = TRUE)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{group_by}{(\emph{required}) Group the samples by a variable in the metadata.}

\item{showSD}{Show standard deviation from mean or not. (\emph{default:} \code{TRUE})}

\item{log10_x}{Log10-transform the x-axis or not. Often most variation is observed among the most abundant OTU's, log10-transforming the x-axis will highlight this better. (\emph{default:} \code{TRUE})}
}
\value{
A ggplot2 object.
}
\description{
Generates a rank abundance curve (rank abundance vs cumulative read abundance), optionally with standard deviation from mean intervals.
}
\details{
Currently only OTU level is supported.
}
\examples{
# Load example data
data("AalborgWWTPs")

# Rank abundance plot
amp_rankabundance(AalborgWWTPs, group_by = "Plant", showSD = TRUE, log10_x = TRUE)
}
\seealso{
\code{\link{amp_load}}
}
\author{
Kasper Skytte Andersen \email{ksa@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
