% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_rankabundance.R
\name{amp_rankabundance}
\alias{amp_rankabundance}
\title{Rank abundance plot}
\usage{
amp_rankabundance(data)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{group_by}{Group the samples by a variable in the metadata.}

\item{order_group}{A vector to order the groups by.}

\item{plot_log}{(\emph{logical}) Log10-scale the x-axis. (\emph{default:} \code{FALSE})}

\item{tax_add}{Additional taxonomic level(s) to display, e.g. \code{"Phylum"}. (\emph{default:} \code{"none"})}

\item{tax_aggregate}{The taxonomic level to aggregate the OTUs. (\emph{default:} \code{"Genus"})}

\item{tax_empty}{How to show OTUs without taxonomic information. One of the following:
\itemize{
   \item \code{"remove"}: Remove OTUs without taxonomic information.
   \item \code{"best"}: (\emph{default}) Use the best classification possible. 
   \item \code{"OTU"}: Display the OTU name.
   }}

\item{tax_class}{Converts a specific phylum to class level instead, e.g. \code{"p__Proteobacteria"}.}

\item{raw}{(\emph{logical}) Display raw input instead of converting to percentages. (\emph{default:} \code{FALSE})}

\item{detailed_output}{(\emph{logical}) Return additional details or not. If \code{TRUE}, it is recommended to save to an object and then access the additional data by \code{View(object$data)}. (\emph{default:} \code{FALSE})}
}
\value{
A ggplot2 object. If \code{detailed_output = TRUE} a list with a ggplot2 object and additional data.
}
\description{
Generates a rank abundance curve (rank abundance vs cumulative read abundance).
}
\details{
A rank abundance curve is used to assess the biodiversity by plotting the ranked abundances of the OTUs (rank 1 is the most abundant, rank 2 the second and so on) versus the cumulative read abundance of the particular OTU. The rank abundances of the OTUs can be grouped by any taxonomic level by the \code{tax_aggregate} argument, the default is per OTU. When the samples are grouped by the \code{group_by} argument, the average of the group is used.
}
\examples{
#Load example data
data("AalborgWWTPs")

#Rank abundance plot
amp_rankabundance(AalborgWWTPs)

}
\author{
Kasper Skytte Andersen \email{kasperskytteandersen@gmail.com}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
