% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_rarecurve.R
\name{amp_rarecurve}
\alias{amp_rarecurve}
\title{Calculate rarefaction curve for each sample.}
\usage{
amp_rarecurve(data)
}
\arguments{
\item{data}{(required) A ampvis object.}

\item{step}{Step size for sample sizes in rarefaction curves (default: 1000).}

\item{color}{A metadata variable to color by.}
}
\description{
Calculate rarefaction curve for each sample using the vegan rarecurve function directly from an ampvis object.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
