% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_rarecurve.R
\name{amp_rarecurve}
\alias{amp_rarecurve}
\title{Rarefaction curve}
\usage{
amp_rarecurve(data)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{stepsize}{Step size for the curves. Lower is prettier but takes more time to generate. (\emph{default:} \code{1000})}

\item{color_by}{Color curves by a variable in the metadata.}
}
\value{
A ggplot2 object.
}
\description{
Generates a rarefaction curve (number of reads vs number of observed OTUs) for each sample.
}
\examples{
#Load example data
data("AalborgWWTPs")

#Rarecurve
amp_rarecurve(AalborgWWTPs)

}
\author{
Kasper Skytte Andersen \email{kasperskytteandersen@gmail.com}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
