% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_rename.R
\name{amp_rename}
\alias{amp_rename}
\title{Used for cleaning and renaming taxonomy}
\usage{
amp_rename(data)
}
\arguments{
\item{data}{(required) A ampvis formated list with all data.}

\item{tax.class}{Converts a specific phyla to class level instead (e.g. "p__Proteobacteria").}

\item{tax.empty}{Either "remove" OTUs without taxonomic information at X level, with "best" classification or add the "OTU" name (default: best).}

\item{tax.level}{The taxonomic level to remove OTUs with empty taxonomy, only used when tax.empty = "remove" (default: Genus).}
}
\value{
A phyloseq object with cleaned and renamed taxonomy.
}
\description{
Used internally in other ampvis functions.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
