% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_subset_samples.R
\name{amp_subset_samples}
\alias{amp_subset_samples}
\title{Subset ampvis2 objects based on sample metadata}
\usage{
amp_subset_samples(data, ...)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{amp_load()}.}

\item{...}{(optional) Logical expression indicating elements or rows to keep in the metadata. Missing values are taken as false. Directly passed to \code{subset()}.}

\item{minreads}{(optional) Minimum number of reads pr. sample. (\emph{default:} \code{1})}
}
\value{
A list with 3 dataframes (4 if reference sequences are provided).
}
\description{
Subsets the data in ampvis2 objects based on metadata and returns the subsetted object. See examples.
}
\examples{
data("MiDAS")
MiDAS

MiDASsubset <- amp_subset_samples(MiDAS, Plant \%in\% c("Aalborg West", "Aalborg East"))

MiDASsubset

MiDASsubset2 <- amp_subset_samples(MiDAS,
    Plant \%in\% c("Aalborg West", "Aalborg East") & !SampleID \%in\% c("16SAMP-749"), 
    minreads = 20000)
    
MiDASsubset2
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
