% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_subset_samples.R
\name{amp_subset_samples}
\alias{amp_subset_samples}
\title{Subset data based on metadata.}
\usage{
amp_subset_samples(data, ...)
}
\arguments{
\item{data}{(required) A object.}

\item{...}{(optional) Additional data is passed on to the R subset function. The samples can be subset based on any variable in the associated metadata.}

\item{minreads}{(optional) Minimum number of reads pr. sample (default: 1).}
}
\value{
A phyloseq object.
}
\description{
Subset data based on metadata.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
