% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_subset_samples.R
\name{amp_subset_samples}
\alias{amp_subset_samples}
\title{Subset ampvis2 objects based on sample metadata}
\usage{
amp_subset_samples(data, ...)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{...}{Logical expression indicating elements or rows to keep in the metadata. Missing values are taken as false. Directly passed to \code{subset()}.}

\item{minreads}{Minimum number of reads pr. sample. Samples below this value will be removed. (\emph{default:} \code{0})}

\item{normalise}{(\emph{logical}) Normalise the read abundances to the total amount of reads (percentages) \emph{BEFORE} the subset. (\emph{default:} \code{FALSE})}

\item{removeAbsents}{(\emph{logical}) Whether to remove OTU's that may have 0 read abundance in all samples after the subset. (\emph{default:} \code{TRUE})}
}
\value{
A modifed ampvis2 object
}
\description{
Subsets the data in ampvis2 objects based on metadata and returns the subsetted object.
}
\details{
The subset is performed on the metadata by \code{subset()} and the abundance- and taxonomy tables are then adjusted accordingly.
}
\section{Normalising data for use in heatmaps}{

By default the raw read counts in the abundance matrix are normalised (transformed to percentages) by \code{\link{amp_heatmap}} automatically. This means that the relative abundances shown will be calculated based on the remaining taxa after the subset, not including the removed taxa, if any. To circumvent this, set \code{normalise = TRUE} when subsetting with the \code{\link{amp_subset_taxa}} and \code{\link{amp_subset_samples}} functions and then set \code{raw = TRUE} when using \code{\link{amp_heatmap}}, see the example below.

\preformatted{
data("MiDAS")
subsettedData <- amp_subset_samples(MiDAS,
                                    Plant \%in\% c("Aalborg West", "Aalborg East"),
                                    normalise = TRUE
                                    )
amp_heatmap(subsettedData,
            group_by = "Plant",
            tax_aggregate = "Phylum",
            tax_add = "Genus",
            raw = TRUE
            )
}
}

\examples{
#Load example data
data("MiDAS")

#Show a short summary about the data by simply typing the name of the object in the console
MiDAS

#Keep only samples containing Aalborg West or East in the Plant column
MiDASsubset <- amp_subset_samples(MiDAS, Plant \%in\% c("Aalborg West", "Aalborg East"))

#Summary
MiDASsubset

#Keep only samples containing Aalborg West or East in the Plant column 
#and remove the sample "16SAMP-749". Remove any sample(s) with less than 20000 total reads
MiDASsubset2 <- amp_subset_samples(MiDAS,
    Plant \%in\% c("Aalborg West", "Aalborg East") & !SampleID \%in\% c("16SAMP-749"), 
    minreads = 20000)
    
#Summary
MiDASsubset2

}
\seealso{
\code{\link{amp_subset_taxa}}, \code{\link{amp_heatmap}}
}
\author{
Kasper Skytte Andersen \email{kasperskytteandersen@gmail.com}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
