% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_subset_taxa.R
\name{amp_subset_taxa}
\alias{amp_subset_taxa}
\title{Subset data based on taxonomy.}
\usage{
amp_subset_taxa(data, tax_vector)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{amp_load()}.}

\item{tax_vector}{(required) a vector with taxonomic groups e.g. c("p__Chloroflexi","p__Actinobacteria")}
}
\value{
A list object with 3 dataframes: abund, tax, and metadata.
}
\description{
Subset data based on taxonomy.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
