% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_subset_taxa.R
\name{amp_subset_taxa}
\alias{amp_subset_taxa}
\title{Subset ampvis2 objects based on taxonomy}
\usage{
amp_subset_taxa(data, tax_vector)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{amp_load()}.}

\item{tax_vector}{(required) A vector with taxonomic groups, e.g. \code{c("p__Chloroflexi","p__Actinobacteria")}.}
}
\value{
A list with 3 dataframes (4 if reference sequences are provided).
}
\description{
Subsets the data in ampvis2 objects based on taxonomy and returns the subsetted object. See examples.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
