% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_subset_taxa.R
\name{amp_subset_taxa}
\alias{amp_subset_taxa}
\title{Subset data based on taxonomy.}
\usage{
amp_subset_taxa(data, ...)
}
\arguments{
\item{data}{(required) A object.}

\item{tax_vector}{(required) a vector with taxonomic groups e.g. c("p__Chloroflexi","p__Actinobacteria")}
}
\value{
A list object.
}
\description{
Subset data based on taxonomy.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
