% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_timeseries.R
\name{amp_timeseries}
\alias{amp_timeseries}
\title{Abundance timeseries}
\usage{
amp_timeseries(data, time_variable = "")
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{time_variable}{(required) The name of the column in the metadata containing the time variables, e.g. \code{"Date"}. Must be directly compatible with \code{\link{as.Date()}} and preferably of the form \code{"yyyy-mm-dd"} or \code{"\%Y-\%m-\%d"}.}

\item{group_by}{Group the samples by a variable in the metadata.}

\item{split}{Split the plot into subplots of each taxa. (\emph{default:} \code{FALSE})}

\item{tax_show}{The number of taxa to show, or a vector of taxa names. (\emph{default:} \code{6})}

\item{tax_aggregate}{The taxonomic level to aggregate the OTUs. (\emph{default:} \code{"OTU"})}

\item{tax_add}{Additional taxonomic level(s) to display, e.g. \code{"Phylum"}. (\emph{default:} \code{"none"})}

\item{tax_class}{Converts a specific phylum to class level instead, e.g. \code{"p__Proteobacteria"}.}

\item{tax_empty}{How to show OTUs without taxonomic information. One of the following:
\itemize{
   \item \code{"remove"}: Remove OTUs without taxonomic information.
   \item \code{"best"}: (\emph{default}) Use the best classification possible. 
   \item \code{"OTU"}: Display the OTU name.
   }}

\item{raw}{(\emph{logical}) Display raw input instead of converting to percentages. (\emph{default:} \code{FALSE})}

\item{...}{Additional arguments passed to \code{\link{as.Date}} to make the time_variable compatible with the timeseries plot, fx the \code{format} or \code{tz} arguments, see \code{?as.Date}.}
}
\value{
A ggplot2 object.
}
\description{
Generates a timeseries plot showing relative read abundances over time.
}
\examples{
#Load example data
data("AalborgWWTPs")

#Timeseries of the 6 most abundant OTUs based on the "Date" column
amp_timeseries(AalborgWWTPs, time_variable = "Date")

#Timeseries of the 6 most abundant Genera and split into 6 individual plots
amp_timeseries(AalborgWWTPs, time_variable = "Date", split = TRUE, tax_aggregate = "Genus")

}
\author{
Julie Klessner Thun Pedersen \email{julieklessnerthun@gmail.com}
}
\keyword{timeseries}
