% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_timeseries.R
\name{amp_timeseries}
\alias{amp_timeseries}
\title{Timeseries}
\usage{
amp_timeseries(data)
}
\arguments{
\item{data}{(required) Data list as loaded with `amp_load()`.}

\item{time}{(required) The name of the column containing the time variables, e.g. "Date".}

\item{group}{A variable from the associated sample data to group/split samples by (default: "Samples").}

\item{split}{Split the plot into subplot of each taxa (default: F).}

\item{tax.show}{The number of taxa to show or a vector of taxa names (default: 5).}

\item{tax.aggregate}{The taxonomic level that the data should be aggregated to (default: "OTU")}

\item{tax.add}{Additional taxonomic levels to display for each entry, e.g. "Phylum" (default: "none").}

\item{tax.class}{Converts a specific phyla to class level instead, e.g. "p__Proteobacteria" (default: "none) .}

\item{tax.empty}{Either "remove" OTUs without taxonomic information, add "best" classification or add the "OTU" name (default: "best").}

\item{raw}{Display raw input instead of converting to percentages (default: F)}
}
\value{
A ggplot2 object or a list with the ggplot2 object and associated dataframes.
}
\description{
This function allows you plot relative abundance reads over time in a timeseries plot.
}
\author{
Julie Klessner Thun Pedersen \email{julieklessnerthun@gmail.com}
}
\keyword{timeseries}
