% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_venn.R
\name{amp_venn}
\alias{amp_venn}
\title{Venn diagram of core OTUs}
\usage{
amp_venn(
  data,
  group_by = NULL,
  cut_a = 0.1,
  cut_f = 80,
  text_size = 5,
  normalise = TRUE,
  detailed_output = FALSE
)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{group_by}{Group the data based on a sample metadata variable. \code{amp_venn} currently only supports a maximum of 3 different groups.}

\item{cut_a}{Abundance cutoff in percent. OTU's below this abundance are excluded from the analysis. (\emph{default:} \code{0.1})}

\item{cut_f}{Frequency cutoff in percent. OTU's within the top \code{cut_f} of the reads are considered a "core" OTU. (\emph{default:} \code{80})}

\item{text_size}{Size of the plotted text. (\emph{default:} \code{5})}

\item{normalise}{(\emph{logical}) Normalise the OTU read counts to be in percent per sample. (\emph{default:} \code{TRUE})}

\item{detailed_output}{(\emph{logical}) Return additional details or not. If \code{TRUE}, it is recommended to save to an object and then access the additional data by \code{View(object$data)}. (\emph{default:} \code{FALSE})}
}
\value{
A ggplot2 object.
}
\description{
Calculates the number of "core" OTUs shared by groups given thresholds for how frequent the OTUs should be above a certain abundance. Also returns the average abundance of the OTUs in a particular group.
}
\section{Preserving relative abundances in a subset of larger data}{

See \code{?\link{amp_filter_samples}} or the \href{https://madsalbertsen.github.io/ampvis2/articles/faq.html#preserving-relative-abundances-in-a-subset-of-larger-data}{ampvis2 FAQ}.
}

\examples{
# Load example data
data("AalborgWWTPs")

# Venn diagram grouped by WWTP
amp_venn(AalborgWWTPs, group_by = "Plant")
}
\references{
[Saunders et al, 2016](https://www.nature.com/articles/ismej2015117)
}
\seealso{
\code{\link{amp_load}}
}
\author{
Kasper Skytte Andersen \email{ksa@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
