% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_venn.R
\name{amp_venn}
\alias{amp_venn}
\title{Generates a ggplot2 style venn diagram}
\usage{
amp_venn(data)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{amp_load()}.}

\item{group}{Group the data based on a sample variable.}

\item{cut_a}{Abundance cutoff in percent (default: 0.1).}

\item{cut_f}{Frequency cutoff in percent (default: 80).}

\item{text.size}{Size of the plotted text.}

\item{output}{Either "plot" or "complete" (default: "plot").}

\item{raw}{Display raw input instead of converting to percentages (default: F).}
}
\value{
A ggplot2 object
}
\description{
Calculates the number of "core" OTUs shared by groups given tresholds for how frequent the OTUs should be above a certain abundance. Also returns the average abundance of the OTUs in a particular group.
}
\author{
Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
