% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_venn.R
\name{amp_venn}
\alias{amp_venn}
\title{Venn diagram of core OTUs}
\usage{
amp_venn(data)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{group_by}{Group the data based on a sample variable.}

\item{cut_a}{Abundance cutoff in percent. (\emph{default:} \code{0.1})}

\item{cut_f}{Frequency cutoff in percent. (\emph{default:} \code{80})}

\item{text_size}{Size of the plotted text. (\emph{default:} \code{5})}

\item{raw}{(\emph{logical}) Display raw input instead of converting to percentages. (\emph{default:} \code{FALSE})}

\item{detailed_output}{(\emph{logical}) Return additional details or not. If \code{TRUE}, it is recommended to save to an object and then access the additional data by \code{View(object$data)}. (\emph{default:} \code{FALSE})}
}
\value{
A ggplot2 object.
}
\description{
Calculates the number of "core" OTUs shared by groups given thresholds for how frequent the OTUs should be above a certain abundance. Also returns the average abundance of the OTUs in a particular group.
}
\examples{
#Load example data
data("AalborgWWTPs")

#Venn diagram grouped by WWTP
amp_venn(AalborgWWTPs, group_by = "Plant")

}
\author{
Kasper Skytte Andersen \email{kasperskytteandersen@gmail.com}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
