% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_load.R
\name{findOTUcol}
\alias{findOTUcol}
\title{Detect OTU/ASV column in a data frame}
\usage{
findOTUcol(x, OTUcolname = c("OTU", "ASV", "#OTU ID"))
}
\arguments{
\item{x}{A data frame}

\item{OTUcolname}{Character vector with columns names to search for OTUs/ASVs}
}
\value{
A data frame
}
\description{
Detect OTU/ASV column in a data frame
}
\keyword{internal}
