% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{matchOTUs}
\alias{matchOTUs}
\title{Rename OTUs by exact sequence matches from a FASTA file}
\usage{
matchOTUs(data, fasta, unmatched_prefix = "unmatched", rename_unmatched = TRUE)
}
\arguments{
\item{data}{data (\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{fasta}{Path to a FASTA file or a \code{DNAbin} class object with sequences whose names will be used as OTU names by exact matches (i.e. same length, 100\% sequence identity). (\emph{default:} \code{NULL})}

\item{unmatched_prefix}{Prefix used to name any unmatched sequences in the FASTA file An integer counting from 1 will be appended to this prefix, so for example the 123th unmatched sequence will be named \code{unmatched123}, and so on. (\emph{default:} \code{"unmatched"})}

\item{rename_unmatched}{Whether to rename any unmatched sequences or not. (\emph{default:} \code{TRUE})}
}
\value{
An ampvis2 class object
}
\description{
Renames sequences loaded in an ampvis2 object based on exact matches (100% identity and exact same length) in a FASTA file. This is useful for enabling direct cross-study/cross-dataset comparison of OTU/ASV names. This function is also used internally in \code{amp_merge_ampvis2}.
}
