% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAS.R
\docType{data}
\name{MiDAS}
\alias{MiDAS}
\title{MiDAS example data}
\format{An object of class \code{"ampvis2"}. See \code{\link[ampvis2]{amp_load}}.}
\source{
\href{http://midasfieldguide.org}{MiDAS: Field Guide to the Microbes of Activated Sludge and Anaerobic Digesters}
}
\usage{
data(MiDAS)
}
\description{
The microbial community composition of the activated sludge of 55 different Danish Wastewater Treatment Plants from 2006-2013 as determined by 16S rRNA amplicon sequencing using the Illumina platform.
}
\examples{
data(MiDAS)
MiDAS
}
\references{
McIlroy et al. (2015); MiDAS: the field guide to the microbes of activated sludge, Database, Volume 2015, bav062
(\href{https://doi.org/10.1093/database/bav062}{https://doi.org/10.1093/database/bav062})
}
\keyword{datasets}
