% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.ampvis2}
\alias{print.ampvis2}
\title{Prints ampvis2 object summary (internal function)}
\usage{
\method{print}{ampvis2}(x, ...)
}
\arguments{
\item{x}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{...}{not used}
}
\description{
Prints ampvis2 object summary (internal function)
}
\author{
Kasper Skytte Andersen \email{ksa@bio.aau.dk}
}
\keyword{internal}
