% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ampvis2.R
\name{print.ampvis2}
\alias{print.ampvis2}
\title{Prints ampvis2 object summary. Internal function.}
\usage{
\method{print}{ampvis2}(data)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}
}
\value{
Text output in console
}
\description{
Prints ampvis2 object summary. Internal function.
}
\author{
Kasper Skytte Andersen \email{kasperskytteandersen@gmail.com}
}
