% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.hmfunplot}
\alias{print.hmfunplot}
\title{Prints heatmap+functions plot (internal function)}
\usage{
\method{print}{hmfunplot}(x, ...)
}
\arguments{
\item{x}{plot}

\item{...}{not used}
}
\description{
Prints heatmap+functions plot (internal function)
}
\author{
Kasper Skytte Andersen \email{ksa@bio.aau.dk}
}
