% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{unifrac}
\alias{unifrac}
\title{Calculate weighted or unweighted UniFrac distances. Adopted from fastUniFrac() from phyloseq}
\usage{
unifrac(data, weighted = FALSE, normalise = TRUE, num_threads = 1L)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{\link{amp_load}}.}

\item{weighted}{Calculate weighted or unweighted UniFrac distances.}

\item{normalise}{Should the output be normalised such that values range from 0 to 1 independent of branch length values? Note that unweighted UniFrac is always normalised. (\emph{default:} \code{TRUE})}

\item{num_threads}{The number of threads to be used for calculating UniFrac distances. If set to more than \code{1} then this is set by using \code{\link[doParallel]{registerDoParallel}} (\emph{default:} \code{1})}
}
\value{
A distance matrix of class \code{dist}
}
\description{
Calculate weighted or unweighted UniFrac distances. Adopted from fastUniFrac() from phyloseq
}
